ZOOSH_ECU-main/HarleyECUDump/requirements.txt
# Harley ECU Dump Tool - Dependencies
python-can>=4.0.0

